requirements.txt
pyupbit==0.2.34
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
